Extract `Announcement::Reactions` concern
